+++ pkg/types/types_statuscontainer.go
@@ -9,11 +9,11 @@ import (
 type StatusContainer struct {
 	coreTypes.Status `json:",inline"`
 	Items            []coreTypes.CacheItem `json:"items"`
-	NItems           uint64                `json:"nItems"`
+	NItems           int                   `json:"nItems"`
 	LatestUpdate     string                `json:"latestUpdate"`
-	NFolders         uint64                `json:"nFolders"`
-	NFiles           uint64                `json:"nFiles"`
-	NBytes           int64                 `json:"nBytes"`
+	NFolders         int                   `json:"nFolders"`
+	NFiles           int                   `json:"nFiles"`
+	NBytes           int                   `json:"nBytes"`
 }
 
 func (s *StatusContainer) ShallowCopy() StatusContainer {
@@ -28,11 +28,11 @@ func (s *StatusContainer) ShallowCopy() StatusContainer {
 }
 
 func (s *StatusContainer) Summarize() {
-	s.NItems = uint64(len(s.Items))
+	s.NItems = len(s.Items)
 	s.LatestUpdate = time.Now().Format(time.RFC3339)
 	for _, cache := range s.Items {
-		s.NFolders += cache.NFolders
-		s.NFiles += cache.NFiles
-		s.NBytes += cache.SizeInBytes
+		s.NFolders += int(cache.NFolders)
+		s.NFiles += int(cache.NFiles)
+		s.NBytes += int(cache.SizeInBytes)
 	}
 }
